requirements.txt
numpy==1.19.2
tf-nightly
pillow==8.3.2
